Changed Lua Environment
Removed the baked version for the library I created for this purpose. The new version is no longer a global singleton and must instead be passed to DuelFlow on creation through the new EnvironmentContainer class. Also removed some extra moon sharp code that is either inoperable, included in the afore mentioned library or worse than existing systems.